+++ crates/canic-core/src/dto/blob_storage.rs
@@ -1,0 +1,11 @@
+use crate::dto::prelude::*;
+
+///
+/// CreateCertificateResult
+///
+
+#[derive(CandidType, Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
+pub struct CreateCertificateResult {
+    pub method: String,
+    pub blob_hash: String,
+}